+++ XiaoyaRetriever/IRetriever.cs
@@ -5,6 +5,6 @@ namespace XiaoyaRetriever
 {
     public interface IRetriever
     {
-        IEnumerable<int> Retrieve(SearchExpression expression);
+        IEnumerable<ulong> Retrieve(SearchExpression expression);
     }
 }
